move typed_ix
